\name{TCGAlgg-package}
\alias{TCGAlgg-package}
\alias{TCGAlgg}
\docType{package}
\title{
TCGAlggPackage:  TCGA LGG data.
}
\description{
Acquired 2-10-15
}
\details{
\tabular{ll}{
Package: \tab TCGAlgg\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-02-25\cr
License: \tab LGPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Paul Shannon <pshannon@systemsbiology.org>
Maintainer: Paul Shannon <pshannon@systemsbiology.org>
}
\references{
none yet...
}
\keyword{dynamic}
\seealso{
  NA
}
\examples{
  NA
}
